Broadcasting base_link

+++ bathy_construction/include/bathy_construction/bathymap_constructor.hpp
@@ -60,8 +60,8 @@ private:
     tf2_ros::TransformBroadcaster br_;
     std::vector<geometry_msgs::TransformStamped> pings_tf_;
 
-    Eigen::Isometry3f map_tf_;
-    Eigen::Isometry3f odom_tf_;
+    Eigen::Isometry3d map_tf_;
+    Eigen::Isometry3d odom_tf_;
 
     SubmapsVec maps_gt_;
     SubmapsVec traj_pings_;
@@ -69,6 +69,7 @@ private:
     MultibeamSensor<PointT> vox_oc_;
 
     int ping_num_;
+
 };
